backend/requirements.txt
# Dépendances principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.18.3

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cryptography==41.0.7
bcrypt==4.1.2

# Paiements
requests==2.31.0
stripe==7.8.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# ⬅️ AJOUT: Rate limiting
slowapi==0.1.8

# ⬅️ AJOUT: Utilitaires mathématiques pour précision
boto3==1.34.0  # Pour AWS (optionnel)
python-dateutil==2.8.2
pytz==2023.3.post1
pyjwt==2.8.0

# ⬅️ AJOUT: Pour les logs et monitoring
structlog==23.2.0
python-json-logger==2.0.7
python-dotenv==1.0.1

# Développement
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Email (optionnel)
jinja2==3.1.2
email-validator>=2.0.0

# ⬅️ AJOUT: Pour les tests de performance
locust==2.20.1  # Pour tester le rate limiting